Fix Twig/PHP 5.4 support, add SphinxClient
The sphinx api is strictly speaking a vendor addition, but it's impractical to manage it with Composer, and in any case it needed to be modified a little to work with PHP 5.4 (constructor renamed, namespaced).

Array keys in the search results were being mangled. This is fixed by adding a prefix string to each key. A getKey method returns the correct key for use in array searches.